+++ SwiftUI_Pull_to_Refresh/SceneDelegate.swift
@@ -21,9 +21,9 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use a UIHostingController as window root view controller.
         if let windowScene = scene as? UIWindowScene {
             let window = UIWindow(windowScene: windowScene)
-            window.rootViewController = AppHostingController(
-				rootView: CitiesView()
-			)
+            window.rootViewController =
+//				UIHostingController(rootView: PaddingView())
+				AppHostingController(rootView: CitiesView())
             self.window = window
             window.makeKeyAndVisible()
         }